requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.11.1
django-filter==2.4.0
idna==2.10
Pillow==8.1.2
proxy-checker==0.6
ProxyGrab==0.2.5
psycopg2-binary==2.8.6
pycurl==7.43.0.6
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
